+++ DscResource.AnalyzerRules/DscResource.AnalyzerRules.Helper.psm1
@@ -12,8 +12,8 @@
     .OUTPUTS
         [System.Boolean]
 
-   .NOTES
-        I initially just walked up the AST tree till I hit 
+    .NOTES
+        I initially just walked up the AST tree till I hit
         a TypeDefinitionAst that was a class
 
         But...
@@ -60,7 +60,7 @@ function Test-IsInClass
     if ($Ast -is [System.Management.Automation.Language.NamedAttributeArgumentAst])
     {
         # Parent is an Attribute Ast AND
-        $inAClass = $Ast.Parent -is [System.Management.Automation.Language.AttributeAst] -and 
+        $inAClass = $Ast.Parent -is [System.Management.Automation.Language.AttributeAst] -and
             # Grandparent is a Property Member Ast (This Ast Type ONLY shows up inside a TypeDefinitionAst) AND
             $Ast.Parent.Parent -is [System.Management.Automation.Language.PropertyMemberAst] -and
             # Great Grandparent is a Type Definition Ast AND
@@ -73,7 +73,7 @@ function Test-IsInClass
     {
         # Parent is a Function Definition Ast AND
         $inAClass = $Ast.Parent -is [System.Management.Automation.Language.FunctionDefinitionAst] -and
-            # Grandparent is a Function Member Ast (This Ast Type ONLY shows up inside a TypeDefinitionAst) AND 
+            # Grandparent is a Function Member Ast (This Ast Type ONLY shows up inside a TypeDefinitionAst) AND
             $Ast.Parent.Parent -is [System.Management.Automation.Language.FunctionMemberAst] -and
             # Great Grandparent is a Type Definition Ast AND
             $Ast.Parent.Parent.Parent -is [System.Management.Automation.Language.TypeDefinitionAst] -and
@@ -98,7 +98,7 @@ function Test-IsInClass
     .OUTPUTS
         [System.String[]]
 
-   .NOTES
+    .NOTES
         None
 #>
 function Get-StatementBlockAsRows
@@ -136,7 +136,7 @@ function Get-StatementBlockAsRows
     .OUTPUTS
         [System.Boolean]
 
-   .NOTES
+    .NOTES
         None
 #>
 function Test-StatementOpeningBraceOnSameLine
@@ -178,7 +178,7 @@ function Test-StatementOpeningBraceOnSameLine
     .OUTPUTS
         [System.Boolean]
 
-   .NOTES
+    .NOTES
         None
 #>
 function Test-StatementOpeningBraceIsNotFollowedByNewLine
@@ -220,7 +220,7 @@ function Test-StatementOpeningBraceIsNotFollowedByNewLine
     .OUTPUTS
         [System.Boolean]
 
-   .NOTES
+    .NOTES
         None
 #>
 function Test-StatementOpeningBraceIsFollowedByMoreThanOneNewLine
@@ -246,4 +246,39 @@ function Test-StatementOpeningBraceIsFollowedByMoreThanOneNewLine
     } # if
 
     return $false
+}
+
+<#
+    .SYNOPSIS
+        Helper function for the Measure-*Statement PSScriptAnalyzer rules.
+        Tests if the statement at the beginning of the string contains any
+        upper case letters.
+
+    .PARAMETER StatementBlock
+        The StatementBlock that contains the statement to check contains any
+        upper case letters.
+
+    .EXAMPLE
+        Test-StatementContainsUpperCase -StatementBlock $ScriptBlockAst.Extent
+
+    .OUTPUTS
+        [System.Boolean]
+
+    .NOTES
+        None
+#>
+function Test-StatementContainsUpperCase
+{
+    [CmdletBinding()]
+    [OutputType([System.Boolean])]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
+        [System.String]
+        $StatementBlock
+    )
+
+    $statement = [System.Text.RegularExpressions.Regex]::Match($statementBlock,'^[a-zA-Z]*').Value
+    return ($statement -cne $statement.ToLower())
 }